[SIG-1350] refactored Header into DetailHeader
