+++ src/signals/incident-management/containers/IncidentDetail/components/DetailHeader/components/DownloadButton/style.scss
