+++ src/signals/incident-management/containers/IncidentDetail/components/DetailHeader/index.js
@@ -6,21 +6,21 @@ import DownloadButton from './components/DownloadButton';
 
 import './style.scss';
 
-const Header = ({ incident, baseUrl, onThor, accessToken }) => {
+const DetailHeader = ({ incident, baseUrl, onThor, accessToken }) => {
   const status = incident && incident.status && incident.status.state;
   const canSplit = (status === 'm') && !(incident && incident.parent_id);
   const canThor = ['m', 'i', 'b', 'h', 'send failed', 'reopened'].some((value) => value === status);
   const downloadLink = incident._links && incident._links['sia:pdf'] && incident._links['sia:pdf'].href;
 
   return (
-    <header className="header">
+    <header className="detail-header">
       <div className="row">
         <div className="col-12">
           <Link to={`${baseUrl}/incidents`} className="startagain action" >Terug naar overzicht</Link>
         </div>
 
-        <div className="col-6 header__title align-self-center">Melding {incident.id}</div>
-        <div className="col-6 header__buttons d-flex justify-content-end">
+        <div className="col-6 detail-header__title align-self-center">Melding {incident.id}</div>
+        <div className="col-6 detail-header__buttons d-flex justify-content-end">
           {canSplit ?
             <Link
               to={`${baseUrl}/incident/${incident.id}/split`}
@@ -45,7 +45,7 @@ const Header = ({ incident, baseUrl, onThor, accessToken }) => {
   );
 };
 
-Header.propTypes = {
+DetailHeader.propTypes = {
   incident: PropTypes.object.isRequired,
   baseUrl: PropTypes.string.isRequired,
   accessToken: PropTypes.string.isRequired,
@@ -53,4 +53,4 @@ Header.propTypes = {
   onThor: PropTypes.func.isRequired
 };
 
-export default Header;
+export default DetailHeader;
