+++ src/signals/incident-management/containers/IncidentDetail/components/DetailHeader/index.test.js
