+++ src/signals/incident-management/containers/IncidentDetail/components/DetailHeader/style.scss
@@ -1,6 +1,6 @@
 @import "config.scss";
 
-.header {
+.detail-header {
   border-bottom: 1px solid #e6e6e6;
